# [Performance][Core] Optimize the performance of evictor v1 and v2 by applying a priority queue and lazy deletion