--- a/vllm/core/evictor.py
+++ b/vllm/core/evictor.py
@@ -114,7 +114,8 @@ class LRUEvictor(Evictor):
         self.free_table[block_id].last_accessed = last_accessed
 
     def _cleanup_if_necessary(self):
-        if len(self.priority_queue) > LRUEvictor.CLEANUP_THRESHOLD * len(self.free_table):
+        if len(self.priority_queue) > LRUEvictor.CLEANUP_THRESHOLD * len(
+                self.free_table):
             self._cleanup()
 
     def _cleanup(self):