--- a/vllm/core/evictor.py
+++ b/vllm/core/evictor.py
@@ -3,6 +3,8 @@ import heapq
 from abc import ABC, abstractmethod
 from typing import Dict, List, Tuple
 
+CLEANUP_THRESHOLD = 50
+
 
 class EvictionPolicy(enum.Enum):
     """Enum for eviction policy used by make_evictor to instantiate the correct
@@ -112,7 +114,7 @@ class LRUEvictor(Evictor):
         self.free_table[block_id].last_accessed = last_accessed
 
     def _cleanup_if_necessary(self):
-        if len(self.priority_queue) > 50 * len(self.free_table):
+        if len(self.priority_queue) > CLEANUP_THRESHOLD * len(self.free_table):
             self._cleanup()
 
     def _cleanup(self):