@youkaichao 2024-08-06
thanks for the contribution!

cc @alexm-neuralmagic @cadedaniel for block manager related optimization.