@Yard1 2024-08-06
FYI this PR seems to be optimizing the same path https://github.com/vllm-project/vllm/pull/7193