@alexm-neuralmagic 2024-08-07
btw, I would rename the topic of the PR to "[Performance] ....", since it is not a bugfix