@llsj14 2024-12-11
In my local test, the` test_eviction_alloc_mixed` sometimes passes and sometimes fails.

```
tests/core/block/test_prefix_caching_block.py ................. [  6%]
............................................................... [ 29%]
............................................................... [ 53%]
............................................................... [ 76%]
............................................................... [100%]
=================== 269 passed, 2 warnings in 6.49s ===================
```

I believe the [assertion in this part](https://github.com/vllm-project/vllm/blob/cad5c0a6eda057eeece87a42fff49fef3e18a2ac/tests/core/block/test_prefix_caching_block.py#L580) is not strictly necessary, because all blocks can be candidates for eviction if they have same last accessed time. The key difference is that the previous code search blocks from the beginning of the free table, while my implementation does not. 

@leiwen83 @cadedaniel @comaniac 
Could you check whether it would be fine to remove the assertion mentioned above and review my PR please?